Add ability to dynamically resize canvas.
Some UAs, for example Safari on iOS, have a maximum canvas size. This
value changes depending on the client device's capabilities (and perhaps
other arbitrary magic).

Generally, though most acutely on high-density displays, it's possible
for us to try and produce a canvas which exceeds these limits. The
behaviour in this case seems to be to produce an empty image URI upon
conversion.

This patch detects this condition and retries with a progressively
smaller canvas until either we succeed or the image becomes really small
in at least one dimension.

Adds a 'debug' option to control some assistive console output.